Refactors GlassCard to improve structure
Consolidates GlassCard logic into GlassCardView for better organization and reusability.

Moves extension properties and functions into GlassCardView
from separate extensions, improving code locality.

Removes redundant GlassFavoriteButton and GlassShareButton views, using generic components for greater flexibility.
