--- MacMagazine/Features/MacMagazineLibrary/Sources/MacMagazineLibrary/Extensions/LinearGradient.swift
@@ -1,26 +1,0 @@
-import SwiftUI
-
-public extension LinearGradient {
-    static var fallbackBackground: LinearGradient {
-        LinearGradient(
-            gradient: Gradient(stops: [
-                .init(color: .black.opacity(0.40), location: 0.0),
-                .init(color: .black.opacity(0.85), location: 1.0)
-            ]),
-            startPoint: .top,
-            endPoint: .bottom
-        )
-    }
-    
-    static var gradientOverlay: LinearGradient {
-        LinearGradient(
-            gradient: Gradient(stops: [
-                .init(color: .black.opacity(0.0), location: 0.0),
-                .init(color: .black.opacity(0.60), location: 0.4),
-                .init(color: .black.opacity(0.95), location: 1.0)
-            ]),
-            startPoint: .top,
-            endPoint: .bottom
-        )
-    }
-}
