--- MacMagazine/Features/MacMagazineLibrary/Sources/MacMagazineLibrary/Extensions/VideoDB.swift
@@ -1,16 +1,0 @@
-import YouTubeLibrary
-import SwiftUI
-
-public extension VideoDB {
-    var formattedDateShort: String {
-        pubDate.formattedDate(using: "dd/MM/yy")
-    }
-    
-    var viewsText: String {
-        "\(views.formattedBigNumber) visualizações"
-    }
-    
-    var likesText: String {
-        "\(likes.formattedBigNumber) curtidas"
-    }
-}
